requirements.txt
cryptography==2.7
Flask-Admin==1.5.3
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-PyMongo>=2.0.0
Flask-SQLAlchemy==2.4.1
Flask==1.0.2
gunicorn==20.0.4
jsonpickle==1.1
marshmallow==3.1.1
mysqlclient
pycryptodome==3.4.3
pymongo[tls,srv]
python-dateutil
pyyaml==5.3.1
requests==2.21.0
SQLAlchemy==1.3.16
urllib3==1.24.1
werkzeug==0.16.0
WTForms==2.2.1
pytest==6.1.2
cerberus==1.3.2